requirements.txt
bcrypt==4.0.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
python-slugify==8.0.1
requests==2.31.0
tinydb==4.7.1
uuid==1.30